+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Warning message instead of histogram when all values are 0, to avoid confusing display behaviour: https://github.com/riparias/gbif-alert/issues/92
 - Fix a minor, recently introduced display issue: https://github.com/riparias/gbif-alert/issues/255
 - Unused datasets are automatically cleaned up at import time: https://github.com/riparias/gbif-alert/issues/222
+- Better synchronization of the Dataset name with GBIF: https://github.com/riparias/gbif-alert/issues/183
 
 # v1.3.0 (2023-08-30)